optimize API request call
