--- frontend/src/atoms/accept.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const acceptState = atom({
-  key: 'accept-state',
-  default: true
-})
