--- frontend/src/atoms/modal.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const modalState = atom({
-    key: "modal-state",
-    default: "open"
-})
