--- frontend/src/atoms/panne_btn.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const panneBtnState = atom({
-    key: "panne_btn_state",
-    default: 10
-})
