--- frontend/src/atoms/zom_leve.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil"
-
-export const zoomLevelState = atom({
-  key: "zoom_level_state",
-  default: 7
-})
