+++ frontend/src/config.ts
@@ -1,2 +1,3 @@
-export const API_BASE_URL =
-    process.env.REACT_APP_API_BASE_URL || 'http://localhost:5000/api';
+export const config = {
+    apiUrl: process.env.REACT_APP_API_BASE_URL || 'http://localhost:8000/api'
+}
