+++ frontend/src/services/api/index.ts
@@ -1,0 +1,2 @@
+export {default as AlertService} from './AlertService';
+export {default as CityService} from './CityService';
