+++ frontend/src/services/api/urls.ts
@@ -1,0 +1,10 @@
+import {config} from '../../config';
+
+const prefixer = config.apiUrl;
+
+export const globalUrls = {
+    GET_ALERTS: `${prefixer}/alerts`,
+    GET_REGION_ALERTS: (region: string) => `${prefixer}/alerts?region=${region}`,
+    GET_CITIES: `${prefixer}/cities`,
+    GET_REGION: `${prefixer}/regions`,
+}
